+++ bin/playgroundbook
@@ -9,12 +9,12 @@ end
 command = ARGV[0]
 file_name = ARGV[1]
 
+require 'playgroundbook'
+
 if command == 'lint'
-  # require 'playgroundbook_lint'
-  # Playgroundbook::Linter.new(file_name).lint
+  Playgroundbook::Linter.new(file_name).lint
 elsif command == 'render'
-  # require 'playgroundbook_render'
-  # Playgroundbook::Renderer.new(file_name).render
+  Playgroundbook::Renderer.new(file_name).render
 else
   puts "Unknown command: #{command}"
   exit 1
