+++ lib/playgroundbook.rb
@@ -1,0 +1,2 @@
+require 'playgroundbook_lint'
+require 'playgroundbook_renderer'
