+++ lib/playgroundbook_lint/abstract_linter.rb
