+++ lib/playgroundbook_lint/chapter_linter.rb
@@ -1,6 +1,6 @@
 require 'plist'
-require 'playgroundbook/abstract_linter'
-require 'playgroundbook/chapter_manifest_linter'
+require 'playgroundbook_lint/abstract_linter'
+require 'playgroundbook_lint/chapter_manifest_linter'
 
 module Playgroundbook
   PAGES_DIRECTORY_NAME = 'Pages'.freeze
