+++ lib/playgroundbook_lint/chapter_manifest_linter.rb
@@ -1,6 +1,6 @@
-require 'playgroundbook/manifest_linter'
-require 'playgroundbook/page_linter'
-require 'playgroundbook/cutscene_page_linter'
+require 'playgroundbook_lint/manifest_linter'
+require 'playgroundbook_lint/page_linter'
+require 'playgroundbook_lint/cutscene_page_linter'
 
 module Playgroundbook
   # A linter for verifying the contents of a chapter's Manifest.plist
