+++ lib/playgroundbook_lint/contents_linter.rb
@@ -1,5 +1,5 @@
-require 'playgroundbook/abstract_linter'
-require 'playgroundbook/root_manifest_linter'
+require 'playgroundbook_lint/abstract_linter'
+require 'playgroundbook_lint/root_manifest_linter'
 
 module Playgroundbook
   # A linter for verifying the contents directory of a playground book
