+++ lib/playgroundbook_lint/cutscene_page_linter.rb
@@ -1,5 +1,5 @@
-require 'playgroundbook/abstract_linter'
-require 'playgroundbook/cutscene_page_manifest_linter'
+require 'playgroundbook_lint/abstract_linter'
+require 'playgroundbook_lint/cutscene_page_manifest_linter'
 
 module Playgroundbook
   # A linter for verifying cutscene pages
