+++ lib/playgroundbook_lint/cutscene_page_manifest_linter.rb
@@ -1,4 +1,4 @@
-require 'playgroundbook/manifest_linter'
+require 'playgroundbook_lint/manifest_linter'
 
 module Playgroundbook
   # A linter for verifying the contents of a cutscene page's manifest
