+++ lib/playgroundbook_lint/root_manifest_linter.rb
@@ -1,6 +1,6 @@
 require 'plist'
-require 'playgroundbook/manifest_linter'
-require 'playgroundbook/chapter_linter'
+require 'playgroundbook_lint/manifest_linter'
+require 'playgroundbook_lint/chapter_linter'
 
 module Playgroundbook
   # A linter for verifying the contents of a playground book's root manifest
