+++ spec/spec_helper.rb
@@ -9,16 +9,16 @@ $LOAD_PATH.unshift((ROOT + 'lib').to_s)
 $LOAD_PATH.unshift((ROOT + 'spec').to_s)
 
 require 'playgroundbook_lint'
-require 'playgroundbook/abstract_linter'
-require 'playgroundbook/chapter_linter'
-require 'playgroundbook/chapter_manifest_linter'
-require 'playgroundbook/contents_linter'
-require 'playgroundbook/manifest_linter'
-require 'playgroundbook/page_linter'
-require 'playgroundbook/page_manifest_linter'
-require 'playgroundbook/cutscene_page_linter'
-require 'playgroundbook/cutscene_page_manifest_linter'
-require 'playgroundbook/root_manifest_linter'
+require 'playgroundbook_lint/abstract_linter'
+require 'playgroundbook_lint/chapter_linter'
+require 'playgroundbook_lint/chapter_manifest_linter'
+require 'playgroundbook_lint/contents_linter'
+require 'playgroundbook_lint/manifest_linter'
+require 'playgroundbook_lint/page_linter'
+require 'playgroundbook_lint/page_manifest_linter'
+require 'playgroundbook_lint/cutscene_page_linter'
+require 'playgroundbook_lint/cutscene_page_manifest_linter'
+require 'playgroundbook_lint/root_manifest_linter'
 
 RSpec.configure do |config|
   config.color = true
